reference-app/trial/requirements.txt
Flask==1.1.4
MarkupSafe==2.0.1
requests==2.23.0
pylint==2.4.0
jinja2==3.0.2
flask_opentracing==1.1.0
gunicorn==20.0.4
jaeger-client==4.3.0
opentelemetry-api==0.12b0
opentelemetry-sdk==0.12b0
opentelemetry-instrumentation==0.12b0
opentelemetry-instrumentation-flask==0.12b0
opentelemetry-instrumentation-requests==0.12b0
opentelemetry-exporter-jaeger==0.12b0
prometheus-flask-exporter==0.18.4